Adding support for trim adjustment settings and on-the-fly reconfiguration.

+++ Nemesis-Mod/src/commands/FlywheelTrimAdjustmentCommand.h
@@ -13,7 +13,7 @@ class FlywheelTrimAdjustmentCommand : public FlywheelCommand {
         void handleImpl(Packet_t packet) override;
 
         FlywheelMotor getMotorFromPacket(Packet_t packet);
-        int getAdjustmentFromPacket(Packet_t packet);        
+        float getAdjustmentFromPacket(Packet_t packet);        
 };
 
 #endif
